+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/type/UpdateBookedQuantityCommerceInventoryAuditTypeImpl.java
@@ -14,6 +14,8 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 
+import java.math.BigDecimal;
+
 import java.util.Locale;
 import java.util.Map;
 
@@ -47,7 +49,7 @@ public class UpdateBookedQuantityCommerceInventoryAuditTypeImpl
 	}
 
 	@Override
-	public String formatQuantity(int quantity, Locale locale) {
+	public String formatQuantity(BigDecimal quantity, Locale locale) {
 		return _language.format(locale, "set-to-x", quantity);
 	}
