+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/type/UpdateWarehouseItemCommerceInventoryAuditTypeImpl.java
@@ -13,6 +13,8 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 
+import java.math.BigDecimal;
+
 import java.util.Locale;
 import java.util.Map;
 
@@ -40,7 +42,7 @@ public class UpdateWarehouseItemCommerceInventoryAuditTypeImpl
 	}
 
 	@Override
-	public String formatQuantity(int quantity, Locale locale) {
+	public String formatQuantity(BigDecimal quantity, Locale locale) {
 		return _language.format(locale, "set-to-x", quantity);
 	}
